Fix compile.
Former-commit-id: 014984e29b61d031a18edb293c0505e884853c1a

+++ source/ch/cyberduck/ui/controller/ITransferView.cs
@@ -63,6 +63,6 @@ namespace Ch.Cyberduck.Ui.Controller
 
         void PopulateBandwidthList(IList<KeyValuePair<float, string>> throttles);
         void TaskbarOverlayIcon(Icon icon, string text);
-        void UpdateOverallProgressState(double progress, double maximum);
+        void UpdateOverallProgressState(long progress, long maximum);
     }
 }
